* fhandler_procsysvipc.cc (format_procsysvipc_msg): Fix printf format 	string for 64 bit. 	* fhandler_registry.cc (fhandler_registry::readdir): Fix cast. 	* fhandler_serial.cc (ibuf): Avoid compiler warning. 	* include/cygwin/ipc.h: Include stdint.h. 	* include/cygwin/msg.h (msgqnum_t): Define as uint32_t. 	(msglen_t): Ditto. 	(struct msginfo): Convert all long members to int32_t. 	(struct msg_info): Ditto. 	* include/cygwin/sem.h (struct seminfo): Ditto. 	* include/cygwin/shm.h (shmatt_t): Define as uint32_t. 	(struct shmid_ds): Define shm_segsz as uint32_t. 	(struct shminfo): Convert all long members to int32_t.

+++ winsup/cygwin/fhandler_registry.cc
@@ -623,7 +623,7 @@ fhandler_registry::readdir (DIR *dir, dirent *de)
       dir->__handle = open_key (path + 1, KEY_READ, wow64, false);
       if (dir->__handle == INVALID_HANDLE_VALUE)
 	goto out;
-      dir->__d_internal = (unsigned) new __DIR_hash ();
+      dir->__d_internal = (uintptr_t) new __DIR_hash ();
     }
   if (dir->__d_position < SPECIAL_DOT_FILE_COUNT)
     {
